move the decorator code into the dpkt package
