+++ dpkt/aoe.py
@@ -2,7 +2,7 @@
 
 import struct
 import dpkt
-from configs.decorators import deprecated_method_decorator
+from decorators import deprecated_method_decorator
 
 
 class AOE(dpkt.Packet):
