+++ dpkt/ntp.py
@@ -3,7 +3,7 @@
 """Network Time Protocol."""
 
 import dpkt
-from configs.decorators import deprecated_method_decorator
+from decorators import deprecated_method_decorator
 
 # NTP v4
