+++ dpkt/pim.py
@@ -3,7 +3,7 @@
 """Protocol Independent Multicast."""
 
 import dpkt
-from configs.decorators import deprecated_method_decorator
+from decorators import deprecated_method_decorator
 
 
 class PIM(dpkt.Packet):
